+++ backend/src/main/java/com/khu/bbangting/repository/FollowRepository.java
@@ -3,8 +3,11 @@ package com.khu.bbangting.repository;
 import com.khu.bbangting.model.Follow;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface FollowRepository extends JpaRepository<Follow, Long> {
     Optional<Follow> findByStoreIdAndUserId(Long storeId, Long userId);
+
+    List<Follow> findAllByUserId(Long userId);
 }
